Added condition rendering and sercured Users, Services with Permissions and Gates from controllers to views

+++ app/Http/Controllers/ServiceController.php
@@ -6,19 +6,21 @@ use App\Models\Service;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Redirect;
 use Inertia\Inertia;
+use Illuminate\Support\Facades\Gate;
 
 class ServiceController extends Controller
 {
     public function index(Request $request)
     {
-        $query = Service::withCount('users'); // Add this line
+        if (!Gate::allows('is_admin')) {
+            abort(403, 'Administratreur seulement');
+        }
+        $query = Service::withCount('users'); 
 
-        // Search by name
         if ($request->has('search')) {
             $query->where('name', 'like', "%{$request->input('search')}%");
         }
 
-        // Sorting
         $query->orderBy($request->sort_by ?? 'name', $request->sort_order ?? 'asc');
 
         $services = $query->paginate(10);
@@ -31,10 +33,17 @@ class ServiceController extends Controller
 
     public function create()
     {
+        if (!Gate::allows('is_admin')) {
+            abort(403, 'Administratreur seulement');
+        }
         return Inertia::render('Services/ServiceCreate');
     }
     public function store(Request $request)
     {
+        if (!Gate::allows('is_admin')) {
+            abort(403, 'Administratreur seulement');
+        }
+        
         $request->validate([
             'name' => 'required|string|max:255|unique:services',
             'description' => 'required|string|max:255',
@@ -50,6 +59,9 @@ class ServiceController extends Controller
 
     public function edit(Service $service)
     {
+        if (!Gate::allows('is_admin')) {
+            abort(403, 'Administratreur seulement');
+        }
         return Inertia::render('Services/ServiceEdit', [
             'service' => $service,
         ]);
@@ -57,6 +69,9 @@ class ServiceController extends Controller
 
     public function update(Request $request, Service $service)
     {
+        if (!Gate::allows('is_admin')) {
+            abort(403, 'Administratreur seulement');
+        } 
         $request->validate([
             'name' => 'required|string|max:255|unique:services,name,' . $service->id,
             'description' => 'nullable|string',
@@ -71,6 +86,9 @@ class ServiceController extends Controller
     }
     public function destroy(Service $service)
     {
+        if (!Gate::allows('is_admin')) {
+            abort(403, 'Administratreur seulement');
+        }
         $service->delete();
         return Redirect::route('services.index')->with('success', 'Service deleted successfully!');
     }
